Fix engine restart
